+++ databricks/sdk/service/ml.py
@@ -882,8 +882,7 @@ class ExperimentAccessControlRequest:
     """Permission level"""
 
     service_principal_name: Optional[str] = None
-    """Application ID of an active service principal. Setting this field requires the
-    `servicePrincipal/user` role."""
+    """application ID of a service principal"""
 
     user_name: Optional[str] = None
     """name of the user"""
@@ -2085,8 +2084,7 @@ class RegisteredModelAccessControlRequest:
     """Permission level"""
 
     service_principal_name: Optional[str] = None
-    """Application ID of an active service principal. Setting this field requires the
-    `servicePrincipal/user` role."""
+    """application ID of a service principal"""
 
     user_name: Optional[str] = None
     """name of the user"""
@@ -3594,7 +3592,8 @@ class ExperimentsAPI:
         """Delete runs by creation time.
         
         Bulk delete runs in an experiment that were created prior to or at the specified timestamp. Deletes at
-        most max_runs per request.
+        most max_runs per request. To call this API from a Databricks Notebook in Python, you can use the
+        client code snippet on https://learn.microsoft.com/en-us/azure/databricks/mlflow/runs#bulk-delete.
         
         :param experiment_id: str
           The ID of the experiment containing the runs to delete.
@@ -4068,7 +4067,8 @@ class ExperimentsAPI:
         """Restore runs by deletion time.
         
         Bulk restore runs in an experiment that were deleted no earlier than the specified timestamp. Restores
-        at most max_runs per request.
+        at most max_runs per request. To call this API from a Databricks Notebook in Python, you can use the
+        client code snippet on https://learn.microsoft.com/en-us/azure/databricks/mlflow/runs#bulk-restore.
         
         :param experiment_id: str
           The ID of the experiment containing the runs to restore.